more clean up, moved more code around
