--- api/src/services/submission-service.ts
@@ -1,3 +1,0 @@
-import { DBService } from "./service";
-
-export class SubmissionService extends DBService { }
